:recycle: LOVE-16 refactor(sign-in-button): Move onboarding components to views folder
