+++ pride/App.js
@@ -5,8 +5,8 @@ import { StatusBar } from 'expo-status-bar';
 
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
-import HomeScreen from './components/HomeScreen';
-import Onboarding from './components/Onboarding';
+import Onboarding from './views/Onboarding';
+import HomeScreen from './views/Onboarding/components/HomeScreen';
 
 import { MainColours, MainStyles } from './styles/core';
