+++ pride/views/Onboarding/components/HomeScreen.js
@@ -1,7 +1,7 @@
 import { View, Text, TouchableOpacity } from 'react-native';
 
 import AsyncStorage from '@react-native-async-storage/async-storage'
-import { MainStyles } from '../styles/core';
+import { MainStyles } from '../../../styles/core';
 
 const HomeScreen = () => {
   const clearOnBoarding = async () => {
